+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog for dets
 
+## 0.11.0
+
+- Updated dependencies
+- Improved the pipeline
+- Support for rest type nodes
+- Support for array literals
+- Renamed `master` to `main`
+- Added `retrieveTypings` API
+- Added guards to support older versions of TypeScript
+
 ## 0.10.2
 
 - Updated dependencies